fixing all bugs in the code

+++ src/domain/comments-service.ts
@@ -2,8 +2,9 @@ import {CommentDbModel, CommentOutputModel} from "../models/comments/comment-mod
 import {commentsRepository} from "../repositories/comments-repository";
 import {OutputPostModel} from "../models/posts/posts-models";
 import {PostsQueryRepository} from "../repositories/posts-query-repository";
+import {commentsQueryRepository} from "../repositories/comments-query-repository";
 
-export class commentsService {
+export class CommentsService {
 
     static async CreateComment(
         userData: {userId: string, userLogin: string}, postId: string, content: string):
@@ -14,6 +15,7 @@ export class commentsService {
 
         if (!post) return null
         const newComment: CommentDbModel = {
+            postId: post.id,
             content: content,
             commentatorInfo: {
                 userId: userData.userId,
@@ -25,12 +27,21 @@ export class commentsService {
         return await commentsRepository.createComment(newComment)
     }
 
-    static async UpdateComment(id: string, body: CommentDbModel) {
-        return await commentsRepository.updateComment(id, body)
+    static async UpdateComment(id: string, body: CommentDbModel, userId: string) {
+     const targetComment = await commentsQueryRepository.getCommentById(id)
+        if (!targetComment) return null;
+        if (targetComment.commentatorInfo.userId != userId) return false;
+         await commentsRepository.updateComment(id, body)
+        return true;
     }
 
-    static async DeleteCommentById(id: string) {
-        return await commentsRepository.deleteComment(id)
+    static async DeleteCommentById(id: string, userId: string) {
+        const targetComment = await commentsQueryRepository.getCommentById(id)
+        if (!targetComment) return null;
+        if (targetComment.commentatorInfo.userId != userId) return false;
+
+        const result= await commentsRepository.deleteComment(id)
+        return result
     }
 
 }
